+++ message_ix_models/tests/util/test_context.py
@@ -8,7 +8,6 @@ import ixmp
 import pytest
 
 from message_ix_models import Context
-from message_ix_models.util import MESSAGE_DATA_PATH
 
 
 class TestContext:
@@ -177,43 +176,3 @@ class TestContext:
             c.use_defaults(defaults)
 
         c.delete()
-
-    # Deprecated methods and attributes
-
-    def test_get_config_file(self, test_context):
-        with pytest.deprecated_call():
-            assert (
-                "message_ix_models",
-                "data",
-                "level.yaml",
-            ) == test_context.get_config_file("level").parts[-3:]
-
-    @pytest.mark.xfail(
-        condition=MESSAGE_DATA_PATH is None,
-        reason="Requires message_data to be installed.",
-    )
-    def test_get_path(self, test_context):
-        with pytest.deprecated_call():
-            assert MESSAGE_DATA_PATH.joinpath(
-                "data", "foo", "bar"
-            ) == test_context.get_path("foo", "bar")
-
-    def test_load_config(self, test_context):
-        # Calling this method is deprecated
-        with pytest.deprecated_call():
-            # Config files can be loaded and are parsed from YAML into Python objects
-            assert isinstance(test_context.load_config("level"), dict)
-
-        # The loaded file is stored and can be reused
-        assert isinstance(test_context["level"], dict)
-
-    def test_units(self, test_context):
-        """Context.units can be used to parse units that are not standard in pint.
-
-        i.e. message_data unit definitions are used.
-        """
-        with pytest.deprecated_call():
-            assert test_context.units("15 USD_2005 / year").dimensionality == {
-                "[currency]": 1,
-                "[time]": -1,
-            }